requirements.txt
PyDAQmx==1.3.2
PyYAML==3.12
numpy==1.12.1
scipy==1.0
Pillow==4.3.0
panda3d==1.10.0.dev1086
